add css, function negative power, change int to float64, make html

+++ mathops/operations.go
@@ -3,24 +3,24 @@ package mathops
 import "errors"
 
 // Add performs addition of two numbers.
-func Add(a, b int) int {
+func Add(a, b float64) float64 {
 	return a + b
 }
 
 // Subtract performs subtraction of two numbers.
-func Subtract(a, b int) int {
+func Subtract(a, b float64) float64 {
 	return a - b
 }
 
 // Multiply performs multiplication of two numbers.
-func Multiply(a, b int) int {
+func Multiply(a, b float64) float64 {
 	return a * b
 }
 
 // Divide performs division of two numbers.
 // It returns both the quotient and remainder.
 // It returns an error if the divisor is zero.
-func Divide(a, b int) (int, error) {
+func Divide(a, b float64) (float64, error) {
 	if b == 0 {
 		return 0, errors.New("division by zero")
 	}
